predicates: add Sebastien's suggestions
