+++ fastjet/internal/predicates/incircle.go
@@ -7,10 +7,6 @@ package predicates
 import (
 	"fmt"
 	"math/big"
-
-	"math"
-
-	"gonum.org/v1/gonum/mat"
 )
 
 // RelativePosition is the position of a point relative to a circle
@@ -174,26 +170,4 @@ func rowBig(a, b, c, d int, plus bool, m []*big.Rat) *big.Rat {
 		bigMul(bigMul(m[a], m[b+4]), bigMul(m[c+8], m[d+12])),
 		bigMul(bigMul(m[b], m[a+4]), bigMul(m[c+8], m[d+12])),
 	)
-}
-
-// matIncircle computes the determinant of the matrix using the mat package.
-// Then it returns the relative position based on the value of the determinant.
-// |1 x1 y1 x1^2+y1^2|
-// |1 x2 y2 x2^2+y2^2|
-// |1 x3 y3 x3^2+y3^2|
-// |1 x  y  x^2 +y^2 |
-func matIncircle(x1, y1, x2, y2, x3, y3, x, y float64) RelativePosition {
-	m := mat.NewDense(4, 4, []float64{1, x1, y1, x1*x1 + y1*y1, 1, x2, y2, x2*x2 + y2*y2, 1, x3, y3, x3*x3 + y3*y3, 1, x, y, x*x + y*y})
-	logDet, sign := mat.LogDet(m)
-	if math.IsInf(logDet, -1) {
-		// logDet is negative infinite and therefore the determinant is 0
-		return On
-	}
-	switch sign {
-	case 1:
-		return Outside
-	case -1:
-		return Inside
-	}
-	return IndeterminatePosition
 }
